Extract `src/compile/scale/type.ts`
